Delete core/arch/arm/tee/arch_tee_fs.c
Tested-by: Pascal Brand <pascal.brand@linaro.org> (STM)
